requirements.txt
pandas==2.2.0
numpy==1.26.3
python-dateutil==2.8.2
pyyaml==6.0.1
loguru==0.7.2 
pyarrow==20.0.0
sqlalchemy==2.0.27 